pyproject.toml
[project]
name = "euxfel"
version = "0.2.0+componentlist.20260121"
description = "The EuXFEL lattice repository"
readme = "README.md"
authors = [
  { name = "Sergey Tomin", email = "" },
  { name = "Bianca Veglia", email = "" },
  { name = "Stuart Walker", email = "stuart.walker@desy.de" }
]

requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "fastexcel>=0.18.0",
    "latdraw>=0.3.1",
    "matplotlib>=3.10.7",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "xlrd>=2.0.2",
    "ocelot-collab",
    "ruff>=0.15.1",
]

[project.scripts]
euxfel = "euxfel.cli:main"

[build-system]
requires = ["uv_build>=0.8.20,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "coverage[toml]>=7.13.1",
    "ipdb>=0.13.13",
    "ipython>=9.7.0",
    "isort>=7.0.0",
    "pyflakes>=3.4.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.uv.sources]
ocelot-collab = { git = "https://github.com/ocelot-collab/ocelot.git", rev = "cadf5d69284416c099088b6fec22e8f1d604ad19" }


[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
# pythonpath = ["src/euxfel"]
addopts = """
-q
--strict-config
--strict-markers
--cov=euxfel
--cov-report=term-missing:skip-covered
--cov-report=xml
--cov-report=html
"""

[tool.coverage.run]
branch = true
source = ["src/euxfel"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
